Fix South Carolina "senior exemption" calculation Fixes #3394
