+++ policyengine_us/tests/policy/baseline/gov/states/sc/tax/income/exemptions/intergration.yaml
@@ -1,0 +1,41 @@
+- name: Tax unit with taxsimid 13715 in f21.its.csv and f21.ots.csv
+  absolute_error_margin: 0.01
+  period: 2021
+  input:
+    people:
+      person1:
+        age: 67
+        employment_income: 27_010
+        taxable_interest_income: 5_505
+        taxable_private_pension_income: 6_000
+        social_security_retirement: 11_500
+        ssi: 0  # not in TAXSIM35
+        state_supplement: 0  # not in TAXSIM35
+        wic: 0  # not in TAXSIM35
+      person2:
+        age: 67
+        employment_income: 30_010
+        taxable_interest_income: 5_505
+        taxable_private_pension_income: 6_000
+        social_security_retirement: 11_500
+        ssi: 0  # not in TAXSIM35
+        state_supplement: 0  # not in TAXSIM35
+        wic: 0  # not in TAXSIM35
+    tax_units:
+      tax_unit:
+        members: [person1, person2]
+        premium_tax_credit: 0  # not in TAXSIM35
+        local_income_tax: 0  # not in TAXSIM35
+        state_sales_tax: 0  # not in TAXSIM35
+    spm_units:
+      spm_unit:
+        members: [person1, person2]
+        snap: 0  # not in TAXSIM35
+        tanf: 0  # not in TAXSIM35
+    households:
+      household:
+        members: [person1, person2]
+        state_fips: 45  # SC
+  output:  # expected results from patched TAXSIM35 2023-12-07 version
+    sc_retirement_deduction: 12_000
+    sc_senior_exemption: 30_000 - 12_000
