+++ policyengine_us/tests/policy/baseline/gov/states/sc/tax/income/exemptions/sc_senior_exemption.yaml
@@ -3,83 +3,58 @@
   absolute_error_margin: 0
   input:
     state_code: SC
-    filing_status: JOINT
-    age_head: 65
-    age_spouse: 65
-    sc_retirement_income_deduction_head: 0
-    sc_retirement_income_deduction_spouse: 2_000
-    sc_military_retirement_income_deduction_head: 0
-    sc_military_retirement_income_deduction_spouse: 0
+    age: 65
+    sc_retirement_deduction_indv: 0
+    sc_military_deduction_indv: 2_000
+    is_tax_unit_head_or_spouse: true
   output:
-    # (15,000 - 0 - 0) + (15,000 - 2_000 - 0) = 28,000
-    sc_senior_exemption: 28_000
+    # (15,000 - 2_000 - 0) = 13,000
+    sc_senior_exemption_person: 13_000
 
-- name: If filing jointly, only one person is age 65 or above
+- name: Age ineligible
   period: 2022
   absolute_error_margin: 0
   input:
     state_code: SC
-    filing_status: JOINT
-    age_head: 65
-    age_spouse: 64
-    sc_retirement_income_deduction_head: 5_000
-    sc_retirement_income_deduction_spouse: 2_000
-    sc_military_retirement_income_deduction_head: 0
-    sc_military_retirement_income_deduction_spouse: 0 
+    age: 64
+    sc_retirement_deduction_indv: 0
+    sc_military_deduction_indv: 2_000
+    is_tax_unit_head_or_spouse: true
   output:
-    # 1*15,000 - 5,000 = 10,000
-    sc_senior_exemption: 10_000
+    sc_senior_exemption_person: 0
 
-
-- name: If filing jointly, head and spouse both under age 65, exemption is $0.
+- name: Not head or spouse so ineligible
   period: 2022
   absolute_error_margin: 0
   input:
     state_code: SC
-    filing_status: JOINT
-    age_head: 64
-    age_spouse: 64
-    sc_retirement_income_deduction_head: 5_000
-    sc_retirement_income_deduction_spouse: 0
-    sc_military_retirement_income_deduction_head: 0
-    sc_military_retirement_income_deduction_spouse: 0
+    age: 65
+    sc_retirement_deduction_indv: 0
+    sc_military_deduction_indv: 2_000
+    is_tax_unit_head_or_spouse: false
   output:
-    sc_senior_exemption: 0
+    sc_senior_exemption_person: 0
 
-- name: If filing seperately, head and spouse are both above 65
+- name: Fully reduced
   period: 2022
   absolute_error_margin: 0
   input:
     state_code: SC
-    filing_status: SEPARATE
-    age_head: 65
-    age_spouse: 65
-    sc_retirement_income_deduction_head: 5_000
-    sc_retirement_income_deduction_spouse: 0
-    sc_military_retirement_income_deduction_head: 0
-    sc_military_retirement_income_deduction_spouse: 0
+    age: 65
+    sc_retirement_deduction_indv: 15_000
+    sc_military_deduction_indv: 2_000
+    is_tax_unit_head_or_spouse: true
   output:
-    # 15,000 - 5,000 - 0 = 10,000
-    sc_senior_exemption: 10_000
+    sc_senior_exemption_person: 0
 
-- name: If filing seperately, head and spouse are both above 65, but the retirement income ducion plus military retrement income deduction is greater than $15,000
+- name: No reduction
   period: 2022
   absolute_error_margin: 0
   input:
     state_code: SC
-    filing_status: SEPARATE
-    age_head: 65
-    age_spouse: 65
-    sc_retirement_income_deduction_head: 8_000
-    sc_retirement_income_deduction_spouse: 2_000
-    sc_military_retirement_income_deduction_head: 7_500
-    sc_military_retirement_income_deduction_spouse: 0
+    age: 65
+    sc_retirement_deduction_indv: 0
+    sc_military_deduction_indv: 0
+    is_tax_unit_head_or_spouse: true
   output:
-    # 15,000 - 8,000 - 7,500 < 0 
-    sc_senior_exemption: 0
-  
-
-
-
-
-
+    sc_senior_exemption_person: 15_000
