--- policyengine_us/variables/gov/states/sc/tax/income/exemptions/sc_military_retirement_income_deduction_spouse.py
@@ -1,10 +1,0 @@
-from policyengine_us.model_api import *
-
-
-class sc_military_retirement_income_deduction_spouse(Variable):
-    value_type = float
-    entity = TaxUnit
-    label = "South Carolina military retirement income deduction for spouse"
-    defined_for = StateCode.SC
-    unit = USD
-    definition_period = YEAR
